requirements.txt
jira==3.5.1
python-dotenv==1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
pandas>=2.1.4
atlassian-python-api>=3.41.4 